backend/requirements.txt
psycopg2-binary==2.9.9
stripe==11.9.0
boto3==1.35.91
anthropic==0.40.0